+++ source/common/upstream/cluster_manager_impl.h
@@ -364,7 +364,7 @@ private:
       if (thread_aware_lb_ != nullptr) {
         return thread_aware_lb_->factory();
       } else {
-        return nullptr;
+        return cluster_->loadBalancerFactory();
       }
     }